rag-chatbot/requirements.txt
sentence-transformers>=2.2.0
chromadb>=0.4.0
langchain-text-splitters>=0.0.1
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
